✨  An attempt to determine the link range (is it ours ?) and transform according to it :
The general idea is to be able to determine whenever the link is in our scope, being able to manipulate it through its full path.

MarkdownDocumentWrapper => FindLinks() can now return absolute links. The Transformation needs it because it is the only information it has.
An other solution would be to pass more information, as the LinkInline.

RepositoryDocumentationInfo => We can determine if the target link is part of the system. This needs more tests and then to be applied to the stack level. The design may change.

RepositoryDocumentationReader => Swap parameters, it was a bug.

It is still unclear to me if a transform Func should return an obsolute path or not. That is why there is a "flag".
I have setup the tests to run both ways all time, and I will decide later. From now on, setting the flag to true seems better, because there is no code needed for the caller.
You can clearly see that discomfort in the tests (and the code of course).