⚗️ CreateRelative works regarding tests data :
I commented out the Debug.Assert as you can see. I was confident that the relative path would be handled without.
It could be, but the reason it does not, is because I assume that the sourcePath does not contains dot : "." or "..".
I can handle it but that will be the work for the optimization.
We have 2 choices here imo :
1. Handle as much as possible cases before it become complex. Meaning doing kind of the same code but reversible. What has been done from target can be done for source.
2. Reduce code size, handle everything at once. Easier to write, but may be way harder to read. It will handle everything at the same time. To make this work, naming has to be super good.

I added a proxy to handle all returns values, it is easier like this to ensure what I always want.
It could be handled in unit tests.