♻ Start working on new structure :
MarkdownDocumentWrapper is responsible for its own creation and can handle more useful values, instead of only exposing links.
Thanks to this, we can now access to an OriginPath or a craftedPath at the same time.
The goal is to structure code in a "stack", each element referencing its children (plural) and its parent (no plural).

The Current property should be the resulting link.