Start handling links alongside a virtual root :
I mostly worked on the fact that a transformation is not anymore returning an absolute link.
Instead, we are able to analyze links more smartly and try to act on it depending of the LinkType.
Also we are crafting a link that is closer to the real output link.
The transformations are safer to run in sequence.
It should be idempotent whatever the order.