Test helper provide some valid context and more :
Those context can be used to test a context of course, but also to get a well formed component of any kind.
Since the coupling between all Md* component is important, this will be a big gain of time.

A context has internal stack, that may be changed in the future but I don't see any point to expose it right now.

Fixed a stack has no parent.
MdContextTests has been adapted with new TestBase.

In and Out are now consistent.