Rewrite transformer that target README.md :
MdDocument takes the responsibility because it is the actual source.
It may be done outside of it when the virtual root is handled totally. That would mean the link processed should be mapped to virtual root all time.

For now it is working as intended.

The result in the processing chains remains to be tested.