CreateRelative has robust approach when handling dots :
The new extension methods seems to work in any cases.
I think this is the way of using it, you never can resolve leading dots, but you simply resolve the rest.
Applying it on failing code solved related tests instantly.
Applying elsewhere kept everything working.
The code is so much simpler to read.