Create a virtual root which is the longest common part within all elements of the context :
Create 2 new test helper properties MultiStackContext and MultiStackWithCrossRefContext that both has 2 stacks.
Context initialization is more robust and can register multiple stacks.
Context first create a virtual root that can be used by any component of the chain.
A document is created from an absolute path and is attached to the virtual root.

The virtual root has to be applied to all the chain now. To documents paths and links.