Check link availability:
It is the first asynchronism, therefore it affect the whole project.
As all the links could be checked by batch, it may be useful to do it to increase performance.
At the moment performance is not an issue with a release loglevel.

The user agent is used to prevent false unauthorized on sites like stackoverflow.
Some sites does not accept HEAD request, so there is a GET fallback.

A configuration context should help skip the links checks.