Support for github branches:
The whole idea got reworked. It may change again.
A repository is now bound to a branch.
The same repository could be used with different branches (untested).
MdStack.TransformCrossRepositoryUrl now clearly support different kind of hosts.
This method should be moved else where though.

There is a trick to be able to use a real ".git" as test data. Git submodules are in my opinion taking a step where they should not.
So it is not possible to commit a ".git" as any folder. That being said, the solution is pretty easy to understand and works like a charm.