Huge refactoring and breaking changes:
A lot of changes got triggered because of the new classes to feed the context.

"Current" is now lazy loaded while it let the context getting completely initialized before being resolved.
The context still has to be able to generate a correct virtual root before constructing anything. This should be the next step.
To create a context, the process has changed. Now it rely on a factory, make it easier to manage async state and monitor. More will be transfered to the factory.
The context can now add on the fly more repositories and recompute all "Current".

Code files are set to target remote. To do that, local code files are initialized with a value that target the remote. Then are locked.
Some commented code show a version of this were the code was copied internally. It is not removed yet, in case it may have a potential usage.

MdStack become MdWorld, as it is branded with a version.

MdRepository has a gitRef and a gitBranch. The gitBranch target the actual local branch. It may be changed in the future to get all the branches.